Changed Integer to Long to match Stream API (default Collections.counting())

+++ Project02Eclipse/src/com/hemebiotech/analytics/SymptomsWriter.java
@@ -19,11 +19,11 @@ public class SymptomsWriter implements ISymptomWriter {
      * @param symptomsCount the HashMap with key/value, a pair of symptoms name/counter
      */
     @Override
-    public void writeSymptoms(Map<String, Integer> symptomsCount) {
+    public void writeSymptoms(Map<String, Long> symptomsCount) {
         FileWriter fileWriter = null;
         try {
             fileWriter = new FileWriter(filePath);
-            for (Map.Entry<String, Integer> entry : symptomsCount.entrySet()) {
+            for (Map.Entry<String, Long> entry : symptomsCount.entrySet()) {
                 String newSymptom = entry.getKey() + " = " + entry.getValue() + "\n";
                 fileWriter.write(newSymptom);
             }
